translate the status code to string when doing data-view binding
